stock_analysis/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
arrow==1.3.0
asttokens==3.0.0
attrs==24.3.0
backoff==2.2.1
backports.tarfile==1.2.0
binaryornot==0.4.4
blinker==1.9.0
boto3==1.35.95
botocore==1.35.84
build==1.2.2.post1
canvas==0.2.5
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cookiecutter==2.6.0
cron-converter==1.2.1
cycler==0.12.1
decorator==5.1.1
dnspython==2.7.0
executing==2.1.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
frozenlist==1.5.0
gql==3.5.0
graphql-core==3.2.5
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
ipython==8.30.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.5.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pip-tools==7.4.1
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.4.0
pymongo==4.10.1
pyparsing==3.2.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
RestrictedPython==7.4
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
statsd==4.0.1
tenacity==9.0.0
text-unidecode==1.3
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
